requirements/base.txt
# ansible>=2.8,<2.9.0
# kombu==4.6.10
# simplejson==3.17.0
Markdown==3.2.2
Pillow==7.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==20.1.0  # https://github.com/hynek/argon2_cffi
bcrypt==3.1.7 # Requires libffi-dev and python-dev
# celery==4.4.6
celery==4.4.7  # pyup: < 5.0  # https://github.com/celery/celery
chardet==3.0.4
colorama==0.4.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.9.2
dj-database-url==0.5.0
django-allauth==0.42.0
django-braces==1.14.0
django-celery-beat==2.0.0  # https://github.com/celery/django-celery-beat
django-ckeditor==5.9.0
django-cors-headers==3.4.0
# django-crispy-forms==1.9.1
django-crispy-forms==1.9.2  # https://github.com/django-crispy-forms/django-crispy-forms
django-environ==0.4.5
# django-extensions==2.2.9
django-filter==2.3.0
django-log-request-id==1.5.0
django-model-utils==4.0.0  # https://github.com/jazzband/django-model-utils
django-phonenumber-field[phonenumbers]==4.0.0
django-redis==4.12.1  # https://github.com/jazzband/django-redis
django-rest-swagger==2.2.0
django-reversion==3.0.7
django-xforwardedfor-middleware==2.0
# django==3.0.9  # pyup: < 3.1  # https://www.djangoproject.com/
django>=3.0,<3.1.0  # pyup: < 3.1  # https://www.djangoproject.com/
djangorestframework-jwt==1.11.0
# djangorestframework==3.11.0
djangorestframework==3.11.1  # https://github.com/encode/django-rest-framework
# flower==0.9.4
flower==0.9.5  # https://github.com/mher/flower
graphene-django==2.10.1
graphene==2.1.8
graypy==2.1.0
hiredis==1.1.0  # https://github.com/redis/hiredis-py
jobtastic==2.1.1
jsonfield==3.1.0
lxml==4.5.1
model-mommy==2.0.0
openapi-codec==1.3.2
passlib==1.7.2
progressbar33==2.4
promise==2.3
python-dateutil==2.8.1
python-slugify==4.0.1  # https://github.com/un33k/python-slugify
pytz==2020.1  # https://github.com/stub42/pytz
pyyaml==5.3.1
redis==3.5.3  # https://github.com/andymccurdy/redis-py
requests==2.24.0
# whitenoise==5.1.0
whitenoise==5.2.0  # https://github.com/evansd/whitenoise

bitcoinaddress==0.1.1
